Refactor CrawlerEngine class
